Moved test helper function to separate file
